# mini-munch-backend/generated-openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MiniMunch API",
    "version": "1.0.0",
    "description": "API and Schema for the MiniMunch restaurant ordering system!"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
    "/api/tables/{tableNumber}": {
      "get": {
        "summary": "Retrieve a table by its number",
        "description": "Retrieves the table document which includes the table ID, table number, and availability status, if successful.",
        "tags": [
          "Tables"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tableNumber",
            "required": true,
            "description": "The number of the table to retrieve.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "table"
                  ],
                  "properties": {
                    "table": {
                      "$ref": "#/components/schemas/Table"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid table number."
          },
          "404": {
            "description": "Table does not exist."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
    "/api/tables/{tableNumber}/occupy": {
      "patch": {
        "summary": "Occupy a table by its number",
        "description": "Checks table availability and occupies the table if it is available. If successful, returns the updated table document (consisting of the table ID, table number, and availability status).",
        "tags": [
          "Tables"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tableNumber",
            "required": true,
            "description": "The number of the table to occupy.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table was occupied successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "table"
                  ],
                  "properties": {
                    "table": {
                      "$ref": "#/components/schemas/Table"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid table number."
          },
          "404": {
            "description": "Table does not exist."
          },
          "409": {
            "description": "Table is already occupied."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
    "/api/tables/{tableNumber}/release": {
      "patch": {
        "summary": "Release a table from being occupied by its number",
        "description": "Checks table availability and releases the table if it is occupied. If successful, returns the updated table document (consisting of the table ID, table number, and availability status).",
        "tags": [
          "Tables"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tableNumber",
            "required": true,
            "description": "The number of the table to release.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table was released successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "table"
                  ],
                  "properties": {
                    "table": {
                      "$ref": "#/components/schemas/Table"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid table number."
          },
          "404": {
            "description": "Table does not exist."
          },
          "409": {
            "description": "Table is already available."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
...
    "/api/orders/table/{tableNumber}": {
      "get": {
        "summary": "Get the order for a table",
        "description": "Retrieves the current order document for a table, which includes the table reference, order items, total, status and creation timestamp. Each table can have at most one order while it is occupied.",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tableNumber",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The table order was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "order"
                  ],
                  "properties": {
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid table number."
          },
          "404": {
            "description": "Table or order does not exist."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Table": {
        "type": "object",
        "required": [
          "_id",
          "tableNumber",
          "available"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "tableNumber": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "available": {
            "type": "boolean"
          }
        }
      },
      "OrderTable": {
        "type": "object",
        "required": [
          "_id",
          "tableNumber"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "tableNumber": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "required": [
          "menuItem",
          "name",
          "price",
          "quantity"
        ],
        "properties": {
          "menuItem": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          }
        }
      },
      "Order": {
        "type": "object",
        "required": [
          "_id",
          "table",
          "items",
          "total",
          "status",
          "createdAt"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "table": {
            "$ref": "#/components/schemas/OrderTable"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "PREPARING",
              "READY"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
}